+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * User Page; show the completion badge
 * User and Fiction Page; show the Read/Unread badge
+* Added selectability to Story Text
 
 ### Changed
 ### Removed